backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
django-filter==23.3
Pillow==10.4.0
PyJWT==2.8.0
setuptools==68.2.2
psycopg[binary]==3.2.9
python-dotenv==1.0.0 